# docs/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "WildGuard proposed backend API",
    "version": "1.0.0",
    "description": "Original proposed full-system contract. Only milestone 1 is implemented; use implemented-openapi.json for its current envelope and APIs. Other modules remain future work."
  },
  "servers": [
    {
      "url": "http://localhost:8080/api/v1",
      "description": "Planned full-system API; see implemented-openapi.json for the implemented subset."
    }
  ],
  "security": [
    {
      "bearerAuth": []
    }
  ],
...
  "paths": {
...
    "/reports/{id}/download": {
      "get": {
        "tags": [
          "Reports"
        ],
        "summary": "Download report as PDF",
        "operationId": "get_Reports_28",
        "x-roles": [
          "PARK_MANAGER"
        ],
        "description": "Generate a small PDF from the stored snapshot. Return Content-Disposition attachment; retain snapshot after download failure. First PDF uses summary tables; maps are a later enhancement.",
        "responses": {
          "200": {
            "description": "Successful response",
            "content": {
              "application/pdf": {
                "schema": {
                  "type": "string",
                  "format": "binary"
                },
                "example": "<PDF bytes>"
              }
            },
            "headers": {
              "Content-Disposition": {
                "schema": {
                  "type": "string"
                },
                "example": "attachment; filename=\"WildGuard_2026-08.pdf\""
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/Error400"
          },
          "401": {
            "$ref": "#/components/responses/Error401"
          },
          "403": {
            "$ref": "#/components/responses/Error403"
          },
          "404": {
            "$ref": "#/components/responses/Error404"
          },
          "500": {
            "$ref": "#/components/responses/Error500"
          },
          "503": {
            "$ref": "#/components/responses/Error503"
          }
        },
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            },
            "example": "77777777-7777-4777-8777-777777777777"
          }
        ]
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    },
    "schemas": {
      "ApiError": {
        "type": "object",
        "properties": {
          "timestamp": {
            "type": "string",
            "format": "date-time"
          },
          "status": {
            "type": "integer",
            "minimum": 400,
            "maximum": 599
          },
          "code": {
            "type": "string"
          },
          "message": {
            "type": "string"
          },
          "path": {
            "type": "string"
          },
          "fieldErrors": {
            "type": "object",
            "additionalProperties": {
              "type": "string"
            }
          }
        },
        "required": [
          "timestamp",
          "status",
          "code",
          "message",
          "path"
        ],
        "additionalProperties": false
      },
...
    },
    "responses": {
      "Error400": {
        "description": "Invalid request or validation failed.",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/ApiError"
            },
            "example": {
              "timestamp": "2026-10-08T06:41:00Z",
              "status": 400,
              "code": "VALIDATION_FAILED",
              "message": "Invalid request or validation failed.",
              "path": "/api/v1/example"
            }
          }
        }
      },
      "Error401": {
        "description": "Missing/invalid token, inactive account or invalid login.",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/ApiError"
            },
            "example": {
              "timestamp": "2026-10-08T06:41:00Z",
              "status": 401,
              "code": "UNAUTHORIZED",
              "message": "Missing/invalid token, inactive account or invalid login.",
              "path": "/api/v1/example"
            }
          }
        }
      },
      "Error403": {
        "description": "Role or requested park is forbidden.",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/ApiError"
            },
            "example": {
              "timestamp": "2026-10-08T06:41:00Z",
              "status": 403,
              "code": "FORBIDDEN",
              "message": "Role or requested park is forbidden.",
              "path": "/api/v1/example"
            }
          }
        }
      },
      "Error404": {
        "description": "Resource not found or not accessible.",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/ApiError"
            },
            "example": {
              "timestamp": "2026-10-08T06:41:00Z",
              "status": 404,
              "code": "NOT_FOUND",
              "message": "Resource not found or not accessible.",
              "path": "/api/v1/example"
            }
          }
        }
      },
...
      "Error500": {
        "description": "Unexpected server error; no false success acknowledgement.",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/ApiError"
            },
            "example": {
              "timestamp": "2026-10-08T06:41:00Z",
              "status": 500,
              "code": "INTERNAL_ERROR",
              "message": "Unexpected server error; no false success acknowledgement.",
              "path": "/api/v1/example"
            }
          }
        }
      },
      "Error503": {
        "description": "Database/storage temporarily unavailable; retry later.",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/ApiError"
            },
            "example": {
              "timestamp": "2026-10-08T06:41:00Z",
              "status": 503,
              "code": "SERVICE_UNAVAILABLE",
              "message": "Database/storage temporarily unavailable; retry later.",
              "path": "/api/v1/example"
            }
          }
        }
      }
    }
  }
}